+++ src/api/notification/notification.user.email.test.ts
@@ -1,4 +1,85 @@
 import { vi, describe, it, expect, beforeEach } from 'vitest'
+import { Recipient, TriggerEvent } from '@opencrvs/toolkit/notification'
+import { TriggerEventPayloadPair } from './handler'
+import { createServer } from '../../index'
+
+const recipient: Recipient = {
+  name: {
+    firstname: 'John',
+    surname: 'Doe'
+  },
+  email: 'john.doe@gmail.com',
+  mobile: '+15551234567'
+}
+
+const userNotificationTestData: TriggerEventPayloadPair[] = [
+  {
+    event: 'user-created',
+    payload: {
+      recipient,
+      username: 'j.doe',
+      temporaryPassword: 'TempPass123!'
+    }
+  },
+  {
+    event: 'user-updated',
+    payload: {
+      recipient,
+      oldUsername: 'z.roronoa',
+      newUsername: 'j.doe'
+    }
+  },
+  {
+    event: 'username-reminder',
+    payload: {
+      recipient,
+      username: 'j.doe'
+    }
+  },
+  {
+    event: '2fa',
+    payload: {
+      recipient,
+      code: '102030'
+    }
+  },
+  {
+    event: 'reset-password',
+    payload: {
+      recipient,
+      code: '112233'
+    }
+  },
+  {
+    event: 'reset-password-by-admin',
+    payload: {
+      recipient,
+      temporaryPassword: 'tempPass123',
+      admin: {
+        name: {
+          firstname: 'Kennedy',
+          surname: 'Campbell'
+        },
+        id: 'admin',
+        role: 'NATIONAL_SYSTEM_ADMIN'
+      }
+    }
+  },
+  {
+    event: TriggerEvent.CHANGE_EMAIL_ADDRESS,
+    payload: {
+      recipient,
+      code: '654321'
+    }
+  },
+  {
+    event: TriggerEvent.CHANGE_PHONE_NUMBER,
+    payload: {
+      recipient,
+      code: '123456'
+    }
+  }
+]
 
 vi.mock('node-fetch', () => {
   return {
@@ -18,10 +99,6 @@ vi.mock('nodemailer', () => {
     }))
   }
 })
-
-import { createServer } from '../../index'
-
-import { userNotificationTestData } from './testData'
 
 describe('User notification - Email', () => {
   let server: any
